refactor: 상수, 타입, 유틸 파일 정리
- validations 유틸 도메인 별로 나눈 파일 모두 관심사별로 통합
