+++ src/components/CardForm/CardFormFields/CardCompanySelect/CardCompanySelect.tsx
@@ -2,14 +2,13 @@ import { ChangeEvent } from "react";
 import {
   CARD_COMPANY_INFO,
   CARD_FORM_TYPE,
-} from "../../../../constants/constants";
+  PLACEHOLDER,
+} from "../../../../constants/card";
 import { useCard } from "../../../../hooks/useCard";
+import { useCardValidation } from "../../../../hooks/useCardValidation";
 import { isCardCompanyState } from "../../../../utils/typeGuard";
 import Select from "../../../Common/Select/Select";
 import { CardFormFieldStyles } from "../CardFormFields.styled";
-import { useCardValidation } from "../../../../hooks/useCardValidation";
-
-const PLACEHOLDER = "카드사를 선택해 주세요";
 
 export default function CardCompanySelect() {
   const { cardCompany, updateCardCompany } = useCard();
@@ -28,7 +27,7 @@ export default function CardCompanySelect() {
       <Select
         key={CARD_FORM_TYPE.cardCompany}
         isError={cardCompanyError.hasError}
-        placeholder={PLACEHOLDER}
+        placeholder={PLACEHOLDER.cardCompany}
         options={CARD_COMPANY_INFO}
         value={cardCompany}
         onChange={handleCardCompanyChange}
