+++ src/components/CardForm/CardFormFields/CardNumberInput/CardNumberInput.tsx
@@ -3,13 +3,13 @@ import {
   CARD_NUMBERS_SEGMENT,
   MAX_LENGTH,
   PLACEHOLDER,
-} from "../../../../constants/constants";
+} from "../../../../constants/card";
 import { useCard } from "../../../../hooks/useCard";
+import { useCardValidation } from "../../../../hooks/useCardValidation";
+import useFieldFocus from "../../../../hooks/useFieldFocus";
+import { CardNumbersSegmentType } from "../../../../types/card";
 import Input from "../../../Common/Input/Input";
 import { CardFormFieldStyles } from "../CardFormFields.styled";
-import { useCardValidation } from "../../../../hooks/useCardValidation";
-import { CardNumbersSegmentType } from "../../../../types/types";
-import useFieldFocus from "../../../../hooks/useFieldFocus";
 
 const cardPositions = Object.values(CARD_NUMBERS_SEGMENT);
