+++ src/constants/validation.ts
@@ -1,0 +1,9 @@
+export const ERROR_MESSAGES = {
+  NUMBER: "숫자만 입력 가능합니다",
+  INVALID_DATE: "유효한 날짜를 입력해 주세요",
+  CHOOSE_CARD_COMPANY: "카드사를 하나 선택해주세요",
+  FILL_CARD_NUMBER: "카드 번호를 모두 입력해 주세요",
+  FILL_DATE: "유효 기간을 모두 입력해 주세요",
+  FILL_CVC: "CVC 번호를 모두 입력해 주세요",
+  FILL_PASSWORD: "비밀번호를 모두 입력해 주세요",
+} as const;
